requirements.txt
matplotlib==3.0.2
numpy==1.15.4
jieba==0.42.1
dataclasses==0.8
six==1.15.0
accelerate==0.4.0
ahocorasick==0.9
datasets==1.12.0
editdistance==0.5.3
espnet_model_zoo==0.1.3
pyarrow==5.0.0
pypinyin==0.42.0
python_Levenshtein==0.12.2
scikit_learn==0.24.2
seqeval==1.2.2
soundfile==0.10.3.post1
torch==1.8.0
transformers==4.5.1
